исправил модули для автономности
