+++ js/big-picture.js
@@ -2,33 +2,29 @@ import {PHOTOS} from './insert.js';
 
 const BIG_PICTURE = document.querySelector('.big-picture');
 const BIG_PICTURE_IMG =  BIG_PICTURE.querySelector('.big-picture__img > img');
-const PICTURES = document.querySelectorAll('.picture');
+
 const LIKES_COUNT = BIG_PICTURE.querySelector('.likes-count');
 const COMMENTS_COUNT = BIG_PICTURE.querySelector('.comments-count');
 const SOCIAL_COMMENT_COUNT = BIG_PICTURE.querySelector('.social__comment-count');
 const COMMENTS_LOADER = BIG_PICTURE.querySelector('.comments-loader');
 const PHOTO_DESCRIPTION = BIG_PICTURE.querySelector('.social__caption');
-
-
-PICTURES.forEach((PICTURE, i) => {
-  PICTURE.addEventListener('click', () => showBigPicture(PICTURE));
-  PICTURE.addEventListener('click', hideComments);
-  PICTURE.addEventListener('click', () => getCommentsList(i));
-  PICTURE.addEventListener('click', () => getPhotoDescription(i));
-  PICTURE.addEventListener('click', spotBodyScroll);
-});
+const BIG_PICTURE_CANCEL = document.querySelector('.big-picture__cancel');
 
 function closeBigPicture() {
   BIG_PICTURE.classList.add('hidden');
   document.body.classList.remove('modal-open');
+  BIG_PICTURE_CANCEL.removeEventListener('click', closeBigPicture);
 }
 
-document.addEventListener('keydown', (event)=> {
-  if (event.code === 'Escape') {
-    BIG_PICTURE.classList.add('hidden');
-    document.body.classList.remove('modal-open');
-  }
-});
+function closeBigPictureByButton () {
+  BIG_PICTURE_CANCEL.addEventListener('click', closeBigPicture);
+
+  document.addEventListener('keydown', (event)=> {
+    if (event.code === 'Escape') {
+      closeBigPicture();
+    }
+  });
+}
 
 function hideComments () {
   SOCIAL_COMMENT_COUNT.classList.add('hidden');
@@ -44,11 +40,11 @@ function showBigPicture (picture) {
   COMMENTS_COUNT.textContent = PICTURE_COMMENTS.textContent;
 }
 
-function getCommentsList (count) {
+function randerCommentsList (index) {
 
   const COMMENT_FRAGMENT = document.createDocumentFragment();
 
-  PHOTOS[count].comments.forEach((comment) => {
+  PHOTOS[index].comments.forEach((comment) => {
     const NEW_COMMENT = BIG_PICTURE.querySelector('.social__comment').cloneNode(true);
     NEW_COMMENT.querySelector('img').src = comment.avatar;
     NEW_COMMENT.querySelector('img').alt = comment.name;
@@ -59,12 +55,12 @@ function getCommentsList (count) {
   BIG_PICTURE.querySelector('.social__comments').append(COMMENT_FRAGMENT);
 }
 
-function getPhotoDescription (count) {
-  PHOTO_DESCRIPTION.textContent = PHOTOS[count].description;
+function setPhotoDescription (index) {
+  PHOTO_DESCRIPTION.textContent = PHOTOS[index].description;
 }
 
-function spotBodyScroll() {
+function stopBodyScroll() {
   document.body.classList.add('modal-open');
 }
 
-export {closeBigPicture};
+export {closeBigPictureByButton, hideComments, showBigPicture, randerCommentsList, setPhotoDescription, stopBodyScroll};
